backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
langchain==0.3.25
langchain-google-genai==0.1.0
google-generativeai==0.4.0
gtts==2.5.4
pillow==10.4.0
requests==2.31.0
